+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Change categories:
 
 - Rework ABI control macros to also enable API control
 - transport: Eliminate poll guards from headers
+- dsp: firmware_update: Move Update Security Revision to ABI Stable
 
 ### Deprecated